fixed minor import issue in settings
