+++ src/AdminPortal/AdminPages/AdminSettings/SettingsPages/ProfileSettings.js
@@ -1,92 +1,132 @@
-import React, { useState } from 'react';
+import React, { useState, useEffect } from "react";
+import { db } from "../../../../firebaseConfig";
+import { doc, getDoc, setDoc } from "firebase/firestore";
 import '../SettingsPagesCSS/ProfileSettings.css';
-import placeholder from "../../../AdminAssets/profileplaceholder.png";
 
 const ProfileSettings = () => {
-  // State Variables for the information
-  const [firstName, setFirstName] = useState('John');
-  const [lastName, setLastName] = useState('Doe');
-  const [email, setEmail] = useState('johndoe@example.com');
-  const [company, setCompany] = useState('Tech Corp');
-  const [bio, setBio] = useState('Software Engineer at Tech Corp');
+    const [firstName, setFirstName] = useState('');
+    const [lastName, setLastName] = useState('');
+    const [email, setEmail] = useState('');
+    const [company, setCompany] = useState('');
+    const [bio, setBio] = useState('');
+    const [icon, setIcon] = useState('');
 
-  // Save button handler
-  const handleProfileChanges = () => {
-    console.log('Profile updated:', { firstName, lastName, email, company, bio});
-  };
+    useEffect(() => {
+        const fetchUserData = async () => {
+            const adminID = "admin2";
+            const docRef = doc(db, "Admins", adminID);
+            const docSnap = await getDoc(docRef);
 
-  return (
-    <div className="ProfileSettings">
-      <div className="profile-container">
-        <h1>User Profile Settings</h1>
-        <div className="profile-main">
+            if(docSnap.exists()) {
+                const data = docSnap.data();
+                setFirstName(data.FirstName);
+                setLastName(data.LastName);
+                setEmail(data.Email);
+                setCompany(data.Company);
+                setBio(data.UserBio);
+                setIcon(data.Icon);
+            }
+            else {
+                console.log("No such document");
+            }
+        };
 
-          {/* Profile Picture Selection */}
-          <div className="profile-picture">
-            <img src={placeholder} alt="Profile" />
-            <div className="display-info">
-              <h2>Display Information</h2>
-              <p><strong>First Name:</strong> {firstName}</p>
-              <p><strong>Last Name:</strong> {lastName}</p>
-              <p><strong>Email:</strong> {email}</p>
-              <p><strong>Company:</strong> {company}</p>
-              <p><strong>Bio:</strong> {bio}</p>
+        fetchUserData();
+    }, []);
+
+    // Save button handler
+    const handleProfileChanges = async () => {
+        const adminID = "admin2";
+        const docRef = doc(db, "Admins", adminID);
+
+        // Update with the new values
+        await setDoc(docRef, {
+            FirstName: firstName,
+            LastName: lastName,
+            Email: email,
+            Company: company,
+            UserBio: bio,
+            Icon: icon
+        }, {merge: true });
+        
+        console.log('Profile updated:', { firstName, lastName, email, company, bio});
+
+        // Show Alert Message
+        alert("Information changed successfully.");
+    };
+
+    return (
+      <div className="ProfileSettings">
+        <div className="profile-info-container">
+          <h1>User Profile Settings</h1>
+          <div className="profile-main">
+
+            {/* Profile Picture Selection */}
+            <div className="profile-picture">
+              <img src={icon} alt="Profile" />
+              <div className="display-info">
+                <h2>Display Information</h2>
+                <p><strong>First Name:</strong> {firstName}</p>
+                <p><strong>Last Name:</strong> {lastName}</p>
+                <p><strong>Email:</strong> {email}</p>
+                <p><strong>Company:</strong> {company}</p>
+                <p><strong>Bio:</strong> {bio}</p>
+              </div>
+            </div>
+
+            {/* Profile Information Section*/}
+            <div className="profile-info">
+              <label>
+                First Name:
+                <input
+                  type="text"
+                  value={firstName}
+                  onChange={(e) => setFirstName(e.target.value)}
+                />
+              </label>
+              <label>
+                Last Name:
+                <input
+                  type="text"
+                  value={lastName}
+                  onChange={(e) => setLastName(e.target.value)}
+                />
+              </label>
+              <label>
+                Email:
+                <input
+                  type="text"
+                  value={email}
+                  onChange={(e) => setEmail(e.target.value)}
+                />
+              </label>
+              <label>
+                Company:
+                <input
+                  type="text"
+                  value={company}
+                  onChange={(e) => setCompany(e.target.value)}
+                />
+              </label>
+              <label>
+                User Bio:
+                <input
+                  type="text"
+                  value={bio}
+                  onChange={(e) => setBio(e.target.value)}
+                  rows="4"
+                />
+              </label>
+
+              {/* Save Changes */}
+              <button onClick={handleProfileChanges}>
+                Save Changes
+              </button>
             </div>
           </div>
-
-          {/* Profile Information Section*/}
-          <div className="profile-info">
-            <label>
-              First Name:
-              <input
-                type="text"
-                value={firstName}
-                onChange={(e) => setFirstName(e.target.value)}
-              />
-            </label>
-            <label>
-              Last Name:
-              <input
-                type="text"
-                value={lastName}
-                onChange={(e) => setLastName(e.target.value)}
-              />
-            </label>
-            <label>
-              Email:
-              <input
-                type="text"
-                value={email}
-                onChange={(e) => setEmail(e.target.value)}
-              />
-            </label>
-            <label>
-              Company:
-              <input
-                type="text"
-                value={company}
-                onChange={(e) => setCompany(e.target.value)}
-              />
-            </label>
-            <label>
-              User Bio:
-              <input
-                type="text"
-                value={bio}
-                onChange={(e) => setBio(e.target.value)}
-                rows="4"
-              />
-            </label>
-
-            {/* Save Changes */}
-            <button onCick={handleProfileChanges}>
-              Save Changes
-            </button>
-          </div>
         </div>
-      </div>
     </div>
-  )
+    );
 };
 
 export default ProfileSettings;
